requirements.txt
Django==1.9.2
django-ipware==1.1.3
c3d==0.3.0
requests==2.9.1
urllib3==1.14
pyenchant==1.6.6
zeroc-ice==3.6.1
django-archive==0.1.5
termcolor
tabulate
ndg-httpsclient
